Creting MysqlGetAllToggleGateway: adding some tests. Updating composer. 70% done.
