+++ test/Toggle/MysqlGetAllTogglesGatewayTest.php
@@ -1,0 +1,163 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Volodymyr
+ * Date: 09/09/2015
+ * Time: 16:12
+ */
+
+namespace Clearbooks\LabsMysql\Toggle;
+
+
+use Clearbooks\Labs\Bootstrap;
+use Clearbooks\Labs\Toggle\Gateway\GetAllTogglesGateway;
+use Clearbooks\LabsMysql\Release\MysqlReleaseGateway;
+use Clearbooks\LabsMysql\Toggle\Entity\Toggle;
+use Doctrine\DBAL\Connection;
+
+class MysqlGetAllTogglesGatewayTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var Connection
+     */
+    private $connection;
+
+    /**
+     * @var GetAllTogglesGateway
+     */
+    private $gateway;
+
+    /**
+     * @param string $releaseName
+     * @param string $url
+     * @return string
+     */
+    private function addRelease( $releaseName, $url )
+    {
+        ( new MysqlReleaseGateway( $this->connection ) )->addRelease( $releaseName, $url );
+        return $this->connection->lastInsertId( "`release`" );
+    }
+
+    /**
+     * @param string $name
+     * @param string $releaseId
+     * @param bool $isActive
+     * @param string $screenshotUrl
+     * @param string $descriptionOfToggle
+     * @param string $descriptionOfFunctionality
+     * @param string $descriptionOfImplementationReason
+     * @param string $descriptionOfLocation
+     * @param string $guideUrl
+     * @param string $appNotificationCopyText
+     * @return string
+     */
+    private function addToggle( $name, $releaseId, $isActive = false, $screenshotUrl = "", $descriptionOfToggle = "",
+                                $descriptionOfFunctionality = "", $descriptionOfImplementationReason = "",
+                                $descriptionOfLocation = "", $guideUrl = "", $appNotificationCopyText = "" )
+    {
+        $this->addToggleToDatabase( $name, $releaseId, $isActive );
+        $toggleId = $this->connection->lastInsertId( "`toggle`" );
+        if (
+            !empty( $screenshotUrl ) ||
+            !empty( $descriptionOfToggle ) ||
+            !empty( $descriptionOfFunctionality ) ||
+            !empty( $descriptionOfImplementationReason ) ||
+            !empty( $descriptionOfLocation ) ||
+            !empty( $guideUrl ) ||
+            !empty( $appNotificationCopyText )
+        ) {
+            $this->addToggleMarketingInformationToDatabase( $toggleId, $screenshotUrl, $descriptionOfToggle,
+                $descriptionOfFunctionality, $descriptionOfImplementationReason,
+                $descriptionOfLocation, $guideUrl, $appNotificationCopyText );
+        }
+        return $toggleId;
+    }
+
+    /**
+     * @param string $name
+     * @param string $releaseId
+     * @param bool $isActive
+     * @return int
+     */
+    public function addToggleToDatabase( $name, $releaseId, $isActive )
+    {
+        return $this->connection->insert( "`toggle`", [
+            'name' => $name,
+            'release_id' => $releaseId,
+            'type' => 1,
+            'visible' => $isActive
+        ] );
+    }
+
+    /**
+     * @param string $toggleId
+     * @param string $screenshotUrl
+     * @param string $descriptionOfToggle
+     * @param string $descriptionOfFunctionality
+     * @param string $descriptionOfImplementationReason
+     * @param string $descriptionOfLocation
+     * @param string $guideUrl
+     * @param string $appNotificationCopyText
+     * @return string
+     */
+    private function addToggleMarketingInformationToDatabase( $toggleId, $screenshotUrl, $descriptionOfToggle,
+                                                              $descriptionOfFunctionality,
+                                                              $descriptionOfImplementationReason,
+                                                              $descriptionOfLocation, $guideUrl,
+                                                              $appNotificationCopyText )
+    {
+        return $this->connection->insert( "`toggle_marketing_information`", [
+            'toggle_id' => $toggleId,
+            'screenshot_urls' => $screenshotUrl,
+            'description_of_toggle' => $descriptionOfToggle,
+            'description_of_functionality' => $descriptionOfFunctionality,
+            'description_of_implementation_reason' => $descriptionOfImplementationReason,
+            'description_of_location' => $descriptionOfLocation, 'guide_url' => $guideUrl,
+            'app_notification_copy_text' => $appNotificationCopyText
+        ] );
+    }
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->connection = Bootstrap::getInstance()->getDIContainer()
+            ->get( Connection::class );
+
+        $this->connection->beginTransaction();
+        $this->connection->setRollbackOnly();
+
+        $this->gateway = new MysqlGetAllTogglesGateway( $this->connection );
+    }
+
+    public function tearDown()
+    {
+        parent::tearDown();
+        $this->connection->rollBack();
+    }
+
+    /**
+     * @test
+     */
+    public function givenNoToggles_ReturnEmptyArray()
+    {
+        $response = $this->gateway->getAllToggles();
+        $this->assertEquals( [ ], $response );
+    }
+
+    /**
+     * @test
+     */
+    public function givenExistentToggles_ReturnArrayOfToggles()
+    {
+        $releaseId = $this->addRelease( "GetAllToggles Release", "Useful url" );
+        $toggleId1 = $this->addToggle( "test", $releaseId, true );
+        $toggleId2 = $this->addToggle( "test2", $releaseId, true );
+
+        $expectedToggles = [ new Toggle( $toggleId1, "test", $releaseId, true ), new Toggle( $toggleId2, "test2",
+            $releaseId, true ) ];
+
+        $response = $this->gateway->getAllToggles();
+        $this->assertEquals( $expectedToggles, $response );
+    }
+}
